Got rid of duplicates FilterField
