+++ ui/src/components/ColumnTitle/CheckboxFilter/CheckboxFilter.less
@@ -21,6 +21,10 @@
     padding-bottom: 5px;
 }
 
+.listItemIcon {
+    margin-right: 4px;
+}
+
 :global(.ant-checkbox-wrapper).checkbox {
     margin-right: 7px;
