+++ ui/src/components/ColumnTitle/FilterField.tsx
@@ -1,25 +1,28 @@
 import React from 'react'
-import { Checkbox } from 'antd'
+import { Checkbox, Icon, Input } from 'antd'
 import { CheckboxChangeEvent } from 'antd/lib/checkbox'
-import { NumberInput, FilterField as CoreFilterField } from '@cxboxComponents'
-import { CheckboxFilter } from './CheckboxFilter/CheckboxFilter'
+import { NumberInput } from '@cxboxComponents'
+import CheckboxFilter from './CheckboxFilter/CheckboxFilter'
 import RangePicker from './RangePicker'
 import DatePicker from './DatePicker'
 import NumberRangeFilter from './components/NumberRangeFilter/NumberRangeFilter'
 import { getFormat } from '@utils/date'
-import { ColumnFilterControlProps } from '@cxboxComponents/ui/FilterField/FilterField'
 import { NumberTypes } from '@cxboxComponents/ui/NumberInput/formaters'
-import { interfaces } from '@cxbox-ui/core'
+import { DataValue, FieldType, RowMetaField, WidgetListField, WidgetMeta } from '@cxbox-ui/core'
 import { DateFieldTypes } from '@interfaces/date'
 import { AppNumberFieldMeta, CustomFieldTypes } from '@interfaces/widget'
 import TimeRangePicker from '@components/ColumnTitle/TimeRangePicker'
 import { ITimePickerFieldMeta } from '../../fields/TimePicker/TimePickerField'
 
-interface FilterFieldProps extends ColumnFilterControlProps {
+interface FilterFieldProps {
+    widgetFieldMeta: WidgetListField
+    rowFieldMeta: RowMetaField
+    value: DataValue | DataValue[]
+    onChange: (value: DataValue | DataValue[]) => void
+    visible?: boolean
+    widgetOptions?: WidgetMeta['options']
     filterByRangeEnabled?: boolean
 }
-
-const { FieldType } = interfaces
 
 function FilterField({ filterByRangeEnabled, ...props }: FilterFieldProps) {
     const { widgetFieldMeta, value, onChange, rowFieldMeta, visible } = props
@@ -47,7 +50,7 @@ function FilterField({ filterByRangeEnabled, ...props }: FilterFieldProps) {
             if (filterByRangeEnabled) {
                 return (
                     <NumberRangeFilter
-                        value={value as interfaces.DataValue[]}
+                        value={value as DataValue[]}
                         type={widgetFieldMeta.type as unknown as NumberTypes}
                         onChange={onChange}
                         digits={fieldMeta.digits}
@@ -74,15 +77,29 @@ function FilterField({ filterByRangeEnabled, ...props }: FilterFieldProps) {
             return (
                 <CheckboxFilter
                     title={widgetFieldMeta.title}
-                    value={value as interfaces.DataValue[]}
+                    value={value as DataValue[]}
                     visible={visible}
                     filterValues={rowFieldMeta.filterValues ? rowFieldMeta.filterValues : []}
                     onChange={onChange}
                 />
             )
         }
-        case FieldType.pickList: {
-            return <CoreFilterField {...props} widgetFieldMeta={{ ...props.widgetFieldMeta, type: FieldType.input }} />
+        case FieldType.dictionary: {
+            const rowFieldMeta = props.rowFieldMeta
+            const rowFieldMetaFilterValues =
+                rowFieldMeta?.filterValues?.map(item => ({
+                    ...item,
+                    icon: rowFieldMeta?.allValues?.find(allValuesItem => allValuesItem.value === item.value)?.icon
+                })) || []
+            return (
+                <CheckboxFilter
+                    title={props.widgetFieldMeta.title}
+                    value={props.value as DataValue[]}
+                    filterValues={rowFieldMetaFilterValues}
+                    visible={props.visible}
+                    onChange={props.onChange}
+                />
+            )
         }
         case FieldType.dateTimeWithSeconds:
         case FieldType.dateTime:
@@ -90,7 +107,7 @@ function FilterField({ filterByRangeEnabled, ...props }: FilterFieldProps) {
             if (filterByRangeEnabled) {
                 return (
                     <RangePicker
-                        value={props.value as interfaces.DataValue[]}
+                        value={props.value as DataValue[]}
                         onChange={onChange}
                         format={getFormat(fieldType === FieldType.dateTime, fieldType === FieldType.dateTimeWithSeconds)}
                         open={visible}
@@ -103,7 +120,7 @@ function FilterField({ filterByRangeEnabled, ...props }: FilterFieldProps) {
                     data-test-filter-popup-value={true}
                     autoFocus
                     onChange={onChange}
-                    value={value as interfaces.DataValue[]}
+                    value={value as DataValue[]}
                     format={getFormat()}
                     open={visible}
                 />
@@ -114,7 +131,7 @@ function FilterField({ filterByRangeEnabled, ...props }: FilterFieldProps) {
             const use12Hours = widgetFieldMeta.format?.includes('A') || widgetFieldMeta.format?.includes('a')
             return (
                 <TimeRangePicker
-                    value={value as interfaces.DataValue[]}
+                    value={value as DataValue[]}
                     onChange={onChange}
                     use12Hours={use12Hours}
                     format={widgetFieldMeta.format}
@@ -124,9 +141,22 @@ function FilterField({ filterByRangeEnabled, ...props }: FilterFieldProps) {
                 />
             )
         }
-
+        case FieldType.pickList:
+        case FieldType.input:
+        case FieldType.text:
         default: {
-            return <CoreFilterField {...props} />
+            return (
+                <Input
+                    data-test-filter-popup-value={true}
+                    autoFocus
+                    value={props.value as string}
+                    suffix={<Icon type="search" />}
+                    onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+                        const textValue = e.target.value.substr(0, 100)
+                        props.onChange(textValue || null)
+                    }}
+                />
+            )
         }
     }
 }
