--- ui/src/cxboxComponents/ui/CheckboxFilter/CheckboxFilter.less
@@ -1,59 +1,0 @@
-/*
- * TESLER-UI
- * Copyright (C) 2018-2020 Tesler Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-.container {
-    width: 100%;
-}
-
-.search {
-    margin-bottom: 8px;
-}
-
-.header {
-    font-weight: bold;
-}
-
-.list {
-    padding: 2px 0;
-    margin-bottom: 0;
-    list-style: none;
-    overflow: auto;
-}
-
-.listItem:not(:last-child) {
-    padding-bottom: 5px;
-}
-
-.listItemIcon {
-    margin-right: 4px;
-}
-
-:global(.ant-checkbox-wrapper).checkbox {
-    margin-right: 7px;
-
-    :global(.ant-checkbox-inner) {
-        border-color: #9c9c9c;
-    }
-}
-
-.noData {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    height: 100%;
-    margin: 0;
-}
