--- ui/src/cxboxComponents/ui/FilterField/FilterField.less
@@ -1,7 +1,0 @@
-.timeFilterPopup {
-   z-index: 1070;
-}
-
-.timeFilter {
-   width: 100%;
-}
