--- ui/src/cxboxComponents/ui/FilterField/components/RangePicker.less
@@ -1,5 +1,0 @@
-.container {
-  :global(.ant-calendar-picker:first-child) {
-    margin-right: 8px;
-  }
-}
